auto_garage/requirements.txt
asgiref==3.10.0
attrs==22.2.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
diff-match-patch==20241021
Django==5.2.7
django-cors-headers==4.9.0
django-import-export==4.3.14
djangorestframework==3.16.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
fonttools==4.38.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
kiwisolver==1.4.4
langcodes==3.3.0
langdetect==1.0.7
langid==1.1.6
MarkupSafe==2.1.2
matplotlib==3.7.0
murmurhash==1.0.9
numpy==1.24.2
packaging==23.0
pathy==0.10.1
Pillow==9.4.0
pluggy==1.0.0
preshed==3.0.8
pydantic==1.10.5
pyparsing==3.0.9
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==1.2.1
requests==2.28.2
six==1.16.0
smart-open==6.3.0
spacy==3.5.0
spacy-langdetect==0.1.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.5.3
srsly==2.4.6
tablib==3.9.0
thinc==8.1.8
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.14
wasabi==1.1.1
